refactor(server): isolate notification delivery routes
